+++ src/utils/cloudinary.ts
@@ -1,4 +1,4 @@
-import { v2 as cloudinary } from 'cloudinary';
+import { ResourceType, v2 as cloudinary } from 'cloudinary';
 import { MFile } from 'types/app';
 import CONFIG from '../core/config';
 
@@ -18,77 +18,160 @@ function extractPathFromCloudinaryURL(url: string): string | null {
   return null;
 }
 
-async function destroyer(url: string) {
+async function destroyer(url: string, type: ResourceType = 'image') {
   const id = extractPathFromCloudinaryURL(url);
 
   if (id) {
-    cloudinary.uploader.destroy(id);
+    await cloudinary.uploader.destroy(id, {
+      resource_type: type,
+    });
   }
 }
 
 const uploadUserAvatar = async (avatar: MFile[], oldAvatar: string) => {
-  if (oldAvatar) {
-    await destroyer(oldAvatar);
+  try {
+    if (oldAvatar) {
+      await destroyer(oldAvatar);
+    }
+
+    const uploadedAvatar = await cloudinary.uploader.upload(avatar[0].path, {
+      folder: `circle/users/avatar`,
+      resource_type: 'image',
+    });
+
+    return uploadedAvatar.secure_url;
+  } catch (error) {
+    return '';
   }
-
-  const uploadedAvatar = await cloudinary.uploader.upload(avatar[0].path, {
-    folder: `circle/users/avatar`,
-    resource_type: 'image',
-  });
-
-  return uploadedAvatar.secure_url;
 };
 
 const uploadUserCover = async (cover: MFile[], oldCover: string) => {
-  if (oldCover) {
-    await destroyer(oldCover);
+  try {
+    if (oldCover) {
+      await destroyer(oldCover);
+    }
+
+    const uploadedCover = await cloudinary.uploader.upload(cover[0].path, {
+      folder: `circle/users/cover`,
+      resource_type: 'image',
+    });
+
+    return uploadedCover.secure_url;
+  } catch (error) {
+    return '';
   }
-
-  const uploadedCover = await cloudinary.uploader.upload(cover[0].path, {
-    folder: `circle/users/cover`,
-    resource_type: 'image',
-  });
-
-  return uploadedCover.secure_url;
 };
 
 const uploadBadgePicture = async (picture: MFile, oldPicture?: string) => {
-  if (oldPicture) {
-    await destroyer(oldPicture);
+  try {
+    if (oldPicture) {
+      await destroyer(oldPicture);
+    }
+
+    const uploadedPicture = await cloudinary.uploader.upload(picture.path, {
+      folder: `circle/badges/pictures`,
+      resource_type: 'image',
+    });
+
+    return uploadedPicture.secure_url;
+  } catch (error) {
+    return '';
   }
-
-  const uploadedPicture = await cloudinary.uploader.upload(picture.path, {
-    folder: `circle/badges/pictures`,
-    resource_type: 'image',
-  });
-
-  return uploadedPicture.secure_url;
 };
 
 const uploadGroupAvatar = async (avatar: MFile[], oldAvatar?: string) => {
-  if (oldAvatar) {
-    await destroyer(oldAvatar);
+  try {
+    if (oldAvatar) {
+      await destroyer(oldAvatar);
+    }
+
+    const uploadedAvatar = await cloudinary.uploader.upload(avatar[0].path, {
+      folder: `circle/groups/avatar`,
+      resource_type: 'image',
+    });
+
+    return uploadedAvatar.secure_url;
+  } catch (error) {
+    return '';
   }
-
-  const uploadedAvatar = await cloudinary.uploader.upload(avatar[0].path, {
-    folder: `circle/groups/avatar`,
-    resource_type: 'image',
-  });
-
-  return uploadedAvatar.secure_url;
 };
 
 const uploadGroupCover = async (cover: MFile[], oldCover?: string) => {
-  if (oldCover) {
-    await destroyer(oldCover);
+  try {
+    if (oldCover) {
+      await destroyer(oldCover);
+    }
+
+    const uploadedCover = await cloudinary.uploader.upload(cover[0].path, {
+      folder: `circle/groups/cover`,
+      resource_type: 'image',
+    });
+
+    return uploadedCover.secure_url;
+  } catch (error) {
+    return '';
   }
+};
 
-  const uploadedCover = await cloudinary.uploader.upload(cover[0].path, {
-    folder: `circle/groups/cover`,
-    resource_type: 'image',
-  });
+const uploadPostVideo = async (video: MFile, oldVideo?: string) => {
+  try {
+    if (oldVideo) {
+      await destroyer(oldVideo, 'video');
+    }
 
-  return uploadedCover.secure_url;
+    const uploadedVideo = await cloudinary.uploader.upload(video.path, {
+      folder: `circle/posts/videos`,
+      resource_type: 'video',
+    });
+
+    return uploadedVideo.secure_url;
+  } catch (error) {
+    return '';
+  }
+};
+
+const uploadPostAudio = async (audio: MFile, oldAudio?: string) => {
+  try {
+    if (oldAudio) {
+      await destroyer(oldAudio, 'video');
+    }
+
+    const uploadedAudio = await cloudinary.uploader.upload(audio.path, {
+      folder: `circle/posts/audios`,
+      resource_type: 'auto',
+    });
+
+    return uploadedAudio.secure_url;
+  } catch (error) {
+    return '';
+  }
+};
+
+const uploadPostGallery = async (
+  pictures: MFile[],
+  oldPictures?: string[]
+): Promise<string[]> => {
+  try {
+    if (oldPictures && oldPictures.length > 0) {
+      await Promise.all(
+        oldPictures.map(async (oldPicture) => await destroyer(oldPicture))
+      );
+    }
+
+    const uploadedPictures = await Promise.all(
+      pictures.map(async (picture) => {
+        const uploadedPicture = await cloudinary.uploader.upload(picture.path, {
+          folder: `circle/posts/gallery`,
+          resource_type: 'image',
+        });
+        return uploadedPicture.secure_url;
+      })
+    );
+
+    return uploadedPictures;
+  } catch (error) {
+    return [];
+  }
 };
 
 export default {
@@ -102,5 +185,10 @@ export default {
   groups: {
     uploadCover: uploadGroupCover,
     uploadAvatar: uploadGroupAvatar,
+  },
+  posts: {
+    uploadVideo: uploadPostVideo,
+    uploadAudio: uploadPostAudio,
+    uploadGallery: uploadPostGallery,
   },
 };