+++ src/utils/multer.ts
@@ -1,5 +1,4 @@
 import multer from 'multer';
-import path from 'path';
 
 // Multer config
 const multerUtils = multer({
@@ -9,16 +8,22 @@ const multerUtils = multer({
   storage: multer.diskStorage({}),
 
   fileFilter: (_request, file, callback) => {
-    const ext = path.extname(file.originalname);
+    const allowedFileTypes = [
+      'image/jpeg',
+      'image/jpg',
+      'image/png',
+      'video/mp4',
+      'audio/mpeg',
+    ];
 
-    if (ext !== '.jpg' && ext !== '.jpeg' && ext !== '.png') {
-      callback(new Error('Only .png, .jpg and .jpeg format allowed!'));
+    if (!allowedFileTypes.includes(file.mimetype)) {
+      callback(new Error('Only mpeg, mp4, jpg, jpeg and png files are allowed!'));
       return;
     }
 
-    // 5 MB
-    if (file.size > 5 * 1024 * 1024) {
-      callback(new Error('File size is too large!'));
+    // 10 MB
+    if (file.size > 10 * 1024 * 1024) {
+      callback(new Error('File size is too large!, max 10MB allowed!'));
       return;
     }
 
@@ -36,5 +41,12 @@ const groupUpload = multerUtils.fields([
   { name: 'cover', maxCount: 1 },
 ]);
 
+const postUpload = multerUtils.fields([
+  { name: 'audio', maxCount: 1 },
+  { name: 'video', maxCount: 1 },
+  { name: 'gallery', maxCount: 10 },
+]);
+
 export default multerUtils;
-export { userUpload, groupUpload };
+export { groupUpload, postUpload, userUpload };
+