+++ CHANGELOG.md
@@ -4,3 +4,10 @@
 ## TODO
 
 - [ ] delete avatar & any images on delete item (user etc)
+- [ ] find a way to handel pagination
+- [ ] add pagination for user bookmarks
+- [ ] change default images & avatars
+- [ ] see post schema and implement it for each collection
+- [ ] create post on update cover or avatar
+- [x] add create post validation
+- [ ] replace search endpoint with all search